More modifications to bring code inline with Rdio TOS
Monitor script now pulls a random track from the queue table's list of identifiers rather than a local copy of the track database. Would prefer to be able to use an Rdio native getRandomTrackInCollection method, or figure out how to access the Rdio API via shell script, but this will do for now.